refactor visibility methods
